Refactored common phone_car code into its own makefile.
The aosp_xxx_car.mk started with definitions that were the same for all of them.
These definititions were moved to a common/pre_google.xml file, and the existing
common/google.xml was renamed to common/post_google.xml

This change also moved the audio_effects.xml file to common/, as they it had the
same content for all devices.

Test: built and booted all phone_car devices (flame, coral, sunfish, blueline, crosshatch, bonito, sargo)
Bug: 160823761

Change-Id: I6f4543b309cb4fda1b04d5ee513d08fc4bd40780